# packages/backend/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "0.1.0",
		"title": "WAFRN Backend",
		"description": "API routes for wafrn social network"
	},
	"paths": {
...
		"/blog/{url}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "url",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}